AutoScoreHighAndAutoBalance
Created a new Autonomous where the robot Scores a cone on the highest pole and then drives backwards for a auto balance.  Needs some more tweaking on on the retracting path as well as the angle for the cone scoring. Then I need it needs to be programmed for it to start climbing the charging station and auto balance.

-Ashley Herrera