StopMotor for Intake
Created a stopMotor command with the help of David for the intake so it can stop after intaking it after a second. Added the Stop motor command as a instant comman in the sequence for auto. This lead to the rest of the code not working for auto, intake didn't intake, nore did it shoot it out to score as well as drive back and start auto balance.